Create bot.py
Main bot script for keeping the bot alive, replying to commands and sending fb posts to subscribed channels.

This is the live code being used in Routiney bot as of now. I know its tooooo messy. I should have used cogs and bot.command, but this is how I started writing the bot so I am continuing on like that. I am not planning to shift codes to cogs and implement any other ease codes as long as the bot is working fine.

Hopefully, I have redacted all the confidential links and token since I just copied and pasted the running bot's code.